requirements.txt
astroid==2.4.2
Babel==2.8.0
bson==0.5.8
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
dataclasses==0.6
defusedxml==0.7.0rc1
Django==2.2.13
django-allauth==0.42.0
django-phonenumber-field==4.0.0
djongo==1.3.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.12.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.10.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.2.4
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
